refactor: rename variadic effects to algebra effects
